requirements.txt
PyMuPDF
python-twitter
python-dotenv
tweepy